# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API Sistem Informasi Akademik Tanya UNPRA
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List jadwal kuliah (semua role)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: prodi
          in: query
          schema: { type: string }
        - name: semester
          in: query
          schema: { type: string }
        - name: tahunAjaran
          in: query
          schema: { type: string }
        - name: lecturerId
          in: query
          schema: { type: string }
        - name: courseId
          in: query
          schema: { type: string }
        - name: hari
          in: query
          schema:
            type: string
            enum: [Senin, Selasa, Rabu, Kamis, Jumat, Sabtu]
      responses:
        "200":
          description: Daftar jadwal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulesResponse"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Tambah jadwal kuliah (admin & dosen)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleRequest"
      responses:
        "201":
          description: Jadwal berhasil dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleResponse"
        "400":
          description: Data tidak valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /schedules/{id}:
    get:
      operationId: getSchedule
      tags: [schedules]
      summary: Detail jadwal kuliah (semua role)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Detail jadwal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleResponse"
        "404":
          description: Jadwal tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update jadwal kuliah (admin & dosen)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScheduleRequest"
      responses:
        "200":
          description: Jadwal berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleResponse"
        "404":
          description: Jadwal tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Hapus jadwal kuliah (admin & dosen)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Jadwal berhasil dihapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Jadwal tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List mahasiswa (admin & dosen)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: prodi
          in: query
          schema: { type: string }
        - name: fakultas
          in: query
          schema: { type: string }
        - name: angkatan
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Daftar mahasiswa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentsResponse"
        "403":
          description: Akses ditolak (mahasiswa tidak bisa list semua)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /students/me:
    get:
      operationId: getMyStudentProfile
      tags: [students]
      summary: Profil mahasiswa yang sedang login (mahasiswa only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profil mahasiswa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentResponse"
        "404":
          description: Data mahasiswa tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateMyStudentProfile
      tags: [students]
      summary: Update profil mahasiswa sendiri (mahasiswa only — hanya phone & address)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMyStudentRequest"
      responses:
        "200":
          description: Profil berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentResponse"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Detail mahasiswa (admin & dosen)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Detail mahasiswa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentResponse"
        "404":
          description: Mahasiswa tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update data mahasiswa (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentRequest"
      responses:
        "200":
          description: Data mahasiswa berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentResponse"
        "404":
          description: Mahasiswa tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Hapus data mahasiswa (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Data mahasiswa berhasil dihapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Mahasiswa tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]
...
    Schedule:
      type: object
      properties:
        id:
          type: string
        courseId:
          type: string
        courseKode:
          type: string
          nullable: true
        courseNama:
          type: string
          nullable: true
        lecturerId:
          type: string
          nullable: true
        lecturerName:
          type: string
          nullable: true
        hari:
          type: string
          enum: [Senin, Selasa, Rabu, Kamis, Jumat, Sabtu]
        jamMulai:
          type: string
          description: Format HH:MM
        jamSelesai:
          type: string
          description: Format HH:MM
        ruangan:
          type: string
        semester:
          type: string
        tahunAjaran:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, courseId, hari, jamMulai, jamSelesai, ruangan, semester, tahunAjaran, createdAt, updatedAt]
    ScheduleResponse:
      type: object
      properties:
        schedule:
          $ref: "#/components/schemas/Schedule"
      required: [schedule]
    SchedulesResponse:
      type: object
      properties:
        schedules:
          type: array
          items:
            $ref: "#/components/schemas/Schedule"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [schedules, pagination]
    CreateScheduleRequest:
      type: object
      properties:
        courseId:
          type: string
          format: uuid
        lecturerId:
          type: string
          format: uuid
        hari:
          type: string
          enum: [Senin, Selasa, Rabu, Kamis, Jumat, Sabtu]
        jamMulai:
          type: string
          description: Format HH:MM (contoh 08:00)
        jamSelesai:
          type: string
          description: Format HH:MM (contoh 10:00)
        ruangan:
          type: string
          maxLength: 50
        semester:
          type: string
        tahunAjaran:
          type: string
      required: [courseId, hari, jamMulai, jamSelesai, ruangan, semester, tahunAjaran]
    UpdateScheduleRequest:
      type: object
      properties:
        courseId:
          type: string
          format: uuid
        lecturerId:
          type: string
          format: uuid
        hari:
          type: string
          enum: [Senin, Selasa, Rabu, Kamis, Jumat, Sabtu]
        jamMulai:
          type: string
        jamSelesai:
          type: string
        ruangan:
          type: string
        semester:
          type: string
        tahunAjaran:
          type: string
...
    Student:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        nim:
          type: string
        prodi:
          type: string
        fakultas:
          type: string
        semester:
          type: integer
        angkatan:
          type: integer
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, nim, prodi, fakultas, semester, angkatan, createdAt, updatedAt]
    StudentResponse:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/Student"
      required: [student]
    StudentsResponse:
      type: object
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [students, pagination]
    UpdateStudentRequest:
      type: object
      properties:
        prodi:
          type: string
        fakultas:
          type: string
        semester:
          type: integer
          minimum: 1
          maximum: 14
        phone:
          type: string
        address:
          type: string
    UpdateMyStudentRequest:
      type: object
      properties:
        phone:
          type: string
          maxLength: 20
        address:
          type: string
          maxLength: 500
...